aicounter/ai/requirements_tes.txt
# ultralytics
argparse
supervision==0.3.1
paho-mqtt==2.0.0
lapx>=0.5.2